testtechnique/requirements.txt
Django>=2.0,<3.0
python-dotenv==0.10.3
pymongo==3.8.0
psycopg2==2.7.7
Pillow==2.2.1